Todas as funcionalidades da primeira parte de lançamentos; Modificações nas bases para cálculos; Novos tipos de bases como INSS e Dependentes foram adicionadas; Modificações Custom JS e Css

+++ codigo-fonte/ImpostoRenda/application/controllers/base.php
@@ -7,11 +7,13 @@ class Base extends CI_Controller {
     public function __construct() {
         parent::__construct();
         $this->load->model("baseModel", "base");
+        $this->load->model("inssModel", "inss");
+        $this->load->model("dependenteModel", "dependente");
     }
 
     public function index(){
         $this->load->view("template/header");
-        $this->load->view("pages/base/listagem", array("list" => $this->base->listAll()));
+        $this->load->view("pages/base/listagem", array("listInss" => $this->inss->listAll(),"listImposto" => $this->base->listAll(),"listDependente" => $this->dependente->listAll()));
         $this->load->view("template/footer");
     }
     
@@ -34,15 +36,14 @@ class Base extends CI_Controller {
     }
     
     // Process
-    
-    
     public function newProcess(){
         
         $datas = array( 
             "valor_inicial" => str_replace(",",".", str_replace(".","",$this->input->post("valor_inicial"))),
             "valor_final" => str_replace(",",".", str_replace(".","",$this->input->post("valor_final"))),
             "aliquota" => $this->input->post("aliquota"),
-            "parcela_deduzir" => str_replace(",",".", str_replace(".","",$this->input->post("deducao")))
+            "parcela_deduzir" => str_replace(",",".", str_replace(".","",$this->input->post("deducao"))),
+            "ano_virgencia" => date("Y")
         );
         
         $response = $this->base->add($datas);
@@ -75,12 +76,34 @@ class Base extends CI_Controller {
     
     public function deleteProcess($id){
         
-        $response = $this->base->delete($id);
-
-        if($response){
-            echo json_encode(array("status" => "success", "message" => "Base deletada com sucesso"));
+        $this->load->model("lancamentoModel", "lancamento");
+        $count = count($this->lancamento->getByBase($id));
+        
+        if(!$count){
+            $response = $this->base->delete($id);
+                if($response){
+                echo json_encode(array("status" => "success", "message" => "Base deletada com sucesso"));
+            }else{
+                echo json_encode(array("status" => "error", "message" => "Ocorreu um problema ao deletar esta base"));
+            }
         }else{
-            echo json_encode(array("status" => "error", "message" => "Ocorreu um problema ao deletar esta base"));
+            
+            if($count == 1){
+                $msg = "01 lançamento foi identificado utilizando esta base. Portanto não será possível remover este registro.";
+            }else{
+                $msg = $count." lançamentos foram identificados utilizando esta base. Portanto não será possível remover este registro.";
+            }
+            
+            echo json_encode(array("status" => "error", "title" => "Atenção", "message" => $msg));
+        }
+        
+    }
+    
+    public function filter($ano = null){
+        if(!is_null($ano)){
+            echo json_encode(array("inss" => $this->inss->listAll($ano),"imposto" => $this->base->listAll($ano)));
+        }else{
+            echo false;
         }
     }
